refactor: add user directory
