+++ src/components/App/App.js
@@ -1,11 +1,11 @@
 import { Route, Routes } from 'react-router-dom';
 import Landing from '../landing/Landing/Landing';
 import Movies from '../movies/Movies/Movies';
-import ProfilePage from '../ProfilePage/ProfilePage';
+import ProfilePage from '../user/ProfilePage/ProfilePage';
 import SavedMovies from '../movies/SavedMovies/SavedMovies';
 import Page404 from '../Page404/Page404';
 import './App.css';
-import Auth from '../auth/Auth/Auth';
+import Auth from '../user/Auth/Auth';
 
 function App() {
   return (
