+++ src/components/user/ProfilePage/ProfilePage.js
@@ -1,5 +1,5 @@
-import Header from '../common/Header/Header';
-import Navigation from '../common/Navigation/Navigation';
+import Header from '../../common/Header/Header';
+import Navigation from '../../common/Navigation/Navigation';
 import Profile from '../Profile/Profile';
 
 function ProfilePage() {
